Add support for Maven wrapper (#1)
https://github.com/takari/maven-wrapper

Maven wrapper ensures that any user of your project can build it, without first worrying about installing Maven. The project comes with scripts that will download and install Maven if it detects that Maven is missing when you run it.

Having the wrapper means we can simplify and make consistent how Maven is installed and managed across all environments; local, CI and wherever else we wish to run it.