Add headless option to config
Both Chrome and Firefox can be driven in headless mode. This means the browser UI is not actually loaded and typically results in faster test times. This is because though the browser will still interact with the web service it doesn't have to render the page.

It should also mean we can actually run our example test in Travis as an extra check that nothing is broken!

N.B. Whilst in we did some housekeeping on the `AcceptanceTestDriver` making private and removing a getter we thought we needed for testing. We also have had to diable a test having spotted it's not actually behaving as expected.