Add the 'test'
This change finally adds the 'test' this project has been built around to better support it.

We have only tweaked the code by dropping some of the opening lines and wrapping it in our own class. For example we no longer need

```java
    // declaration and instantiation of objects/variables
    System.setProperty("webdriver.gecko.driver",
        "C:\\Users\\pmurdock\\Desktop\\MyFiles - not backed up\\geckodriver-v0.21.0-win32\\geckodriver.exe");
    System.setProperty(FirefoxDriver.SystemProperty.DRIVER_USE_MARIONETTE, "true");
    System.setProperty(FirefoxDriver.SystemProperty.BROWSER_LOGFILE, "/dev/null");
    WebDriver driver = new FirefoxDriver();
```

And the first call is now `driver.navigate().to(rootUrl);` instead of the hardcoded versions that existed before, that users were expected to comment out/in to change the environment the 'test' is run against.

Other than that it remains as found 😱