Exclude manifest signature files from uber-jar
https://stackoverflow.com/a/6743609/6117745

In another branch and in an effort to decrypt the password from the 'test' we added a dependency it turns out is signed. When you combine a signed Jar into our uber-jar via the Maven shade plugin the signature no longer matches.

The top answer to resolve this is to exclude manifest signature files by adding a filter to the maven shade configuration. This also means I now know what these lines are doing in the pom.xml!