Fix AcceptanceTestDriver tests for windows
We found when running the unit tests on Windows those tests that were checking whether the `System.getProperty()` call returned the expected string failed. This was because we were testing whether the string ended with `chromedriver` or `geckodriver`. What we had overlooked was that on windows the code will add `.exe` to the end of the string. Hence on Windows, all these acceptance tests fail.

To ensure the tests still checks what we need but also works cross platform, this change replaces `endsWith()` with `contains()` in the tests.