Fix failing ConfigurationTest on Windows
`We_can_instantiate_an_instance_based_on_a_config_file()` passes on unix based systems but not in Windows. The reason is what `File.separator` returns and how that interacts with `classLoader.getResource()`. On unix based systems it returns `/`. `getResource()` does nothing to this hence it works.

On windows `File.separator` returns `\`. In this case `getResource()` encodes it to `%5`. This is because `getResource()` is actually returning a URL not a file path. You could then decode it to get back the `\` and everything works fine. But that becomes a lot of code just so we can put our single file in a folder called **fixtures**.

The simplest solution allround is to just put it at the root of `test/resources` and be done with it!