Fix example test used in Travis CI
The example test is used to check a new installation, and in Travis just to check we can actually run and hit a website. Recently though, the test is failing. After local testing its because the latest version of google chrome is throwing up a dialog about cookies. This appears to be regardless of whether automated testing is running, both normal and headless.

The dialog has broken our test, so we need to implement another, ideally in a page less likely to change than the Google home page.