Fix example test used in Travis CI (#16)
The example test is used to check a new installation, and in Travis just to check we can actually run and hit a website. Recently though, the test is failing. After local testing its because the latest version of google chrome is throwing up a dialog about cookies. This appears to be regardless of whether automated testing is running, both normal and headless.

The dialog has broken our test, so we need to implement another, ideally on a page less likely to change than the Google home page.

We've chosen [example.com](http://example.com) as its such a simple page a lot of folks don't know about that has never changed in all the years I've used it. Hopefully, we should encounter fewer surprises using it.