Grab matching chromedriver version
Our project relies on the chromedriver executable being in a specific location. So using a mix of code from our previous travis config and what we found in the [setup-chromedriver](https://github.com/nanasess/setup-chromedriver) action we have added a step to manually grab the executable and place it in our drivers folder.