Change test user to admin@sroc.gov.uk
Across the environments we found that tacticalcm+1@gmail.com may be confirmed differently. It's also something that has to be manually created. admin@sroc.gov.uk is seeded by the project and has been setup consistenyly across the environments. So to make the test more stable and consistent wherever we run it, this changes the user for the test to one dedicated to acceptance testing.